% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Generate MC simulated data}
\usage{
get_data(x)
}
\arguments{
\item{x}{a data frame with columns nobs, slope, var_ratio, var_r, cv, nsim}
}
\value{
data frame list
}
\description{
Generate MC simulated data
}
\examples{
get_data(data.frame(nobs = 10, slope = 2, var_ratio = .1, var_r = .05, cv = .2, nsim = 10))
}
