% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_plot}
\alias{make_plot}
\title{Create hypothetical power plots}
\usage{
make_plot(df0, frequency, cv, sample_size)
}
\arguments{
\item{df0}{a data frame with columns nobs, slope, var_ratio, var_r, cv, nsim, Power_MK}

\item{frequency}{sampling frequency}

\item{cv}{coefficient of variation for respective frequency}

\item{sample_size}{sample size for respective frequency}
}
\value{
ggplot object
}
\description{
Create hypothetical power plots
}
\examples{
make_plot(data.frame(nobs = c(2, 10), slope = c(2, 2), var_ratio = c(.1, .1), var_r = c(0.05, .05), cv = c(.2, .2), nsim = c(10, 10), Power_MK = c(.1, .311)), "Annually", 0.5, 12)
}
