% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Mode}
\alias{Mode}
\title{Find Mode of a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
mode of the vector
}
\description{
Find Mode of a vector
}
\examples{
Mode(c("cat", "dog", "cat", "fish"))
}
