% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{run tests on nsim datasets}
\usage{
simulate(x)
}
\arguments{
\item{x}{a data frame with columns nobs, slope, var_ratio, var_r, cv, nsim}
}
\value{
Power_MK vector
}
\description{
run tests on nsim datasets
}
\examples{
simulate(data.frame(nobs = 10, slope = 2, var_ratio = .1, var_r = .05, cv = .2, nsim = 10))
}
